pyproject.toml
[project]
name = "habit-tracker-api"
version = "0.1.0"
description = ""
authors = [
    {name = "alnbastos",email = "alinebastos23@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.128.4,<0.129.0)",
    "tortoise-orm[asyncpg] (>=0.25.4,<0.26.0)",
    "aerich[toml] (>=0.9.2,<0.10.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "polyfactory (>=3.3.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = "habit_tracker_api", from = "app"}]


[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.configs.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
